check-config.sh: Remove merge image validation
If we can't trust the 'localhost/apache-base:latest' image to be the parent image of 'localhost/apache:latest', we can't trust that 'localhost/apache:latest' is the image we built... So, this check didn't make much sense from the start. The reason we remove this check now is because '{{.Parent}}' isn't available anymore due to `buildah commit --omit-history`.